+++ MyNPCLib/NPCCommand.cs
@@ -7,7 +7,7 @@ namespace MyNPCLib
     public class NPCCommand : INPCCommand
     {
         public string Name { get; set; }
-        public int InitiatingProcessId { get; set; }
+        public ulong InitiatingProcessId { get; set; }
         public KindOfLinkingToInitiator KindOfLinkingToInitiator { get; set; } = KindOfLinkingToInitiator.Standalone;
         public Dictionary<string, object> Params { get; set; } = new Dictionary<string, object>();
